test: remove calls to clearAllMocks etc
ResizeObserver is mocked as part of test setup, and every time we call a jest method such as
clearAllMocks, resetAllMocks or restoreAllMocks inside a test suite, that applies to the
ResizeObserver mock too. This risks causing later tests that rely on the ResizeObserver mock to
potentially fail. This has been fixed by removing all such calls and replacing them with ones that
apply only to the mock function(s) required by that test suite.