fix(menu): make submenu scrollable when inside a fixed navigation bar
When a submenu is used in a Menu component which is inside a GlobalHeader, or a NavigationBar with
position="fixed", the CSS fixed position means that it is impossible to access all the submenu items
when the total height of the submenu is enough to take it off the screen. (This can happen even with
relatively short submenus if zooming in.) To fix this, a context has been introduced to track when a
submenu is inside such a fixed navigation bar, and CSS is applied to make the submenu scrollable and
with the appropriate max-height.

fix #6009