refactor: make GlobalHeader directly render NavigationBar
The GlobalHeader component is basically a copy of NavigationBar with a different z-index and certain
other CSS properties fixed to particular values. However it isn't actually implemented like that,
instead rendering its own StyledGlobalHeader with those props specified, with this styled component
itself being a copy of StyledNavigationBar but with the z-index overridden. Unfortunately this meant
that the new FixedNavigationBarContext functionality, which uses React Hooks and therefore can't be
put in a styled component, had to be duplicated across the two components - potentially causing
maintenance problems in future. To avoid this, the components have been refactored to allow
GlobalHeader to simply render NavigationBar, and pass it a new private isGlobal prop.