fix(textarea): stop unwanted scroll of container when expandable
When a textarea is `expandable`, it is resized after every render, as well as on resize. This works
by first setting the height to 0, then the appropriate size to contain the text. Setting to 0 first
is necessary in some scenarios, but when the textarea is in a scrollable container, it makes the
container scroll to the top, and then when the height is reset, the browser only scrolls the
container down far enough to reach the cursor position - which may be not as far down as where the
user was previously. The result is an unwanted and unexpected "jump" of the container scroll
position, particularly when typing in a controlled textarea. This is now fixed by taking note of the
previous scroll position and resetting it after resizing is complete.

fix #6088