fix(dialog, sidebar): use aria-hidden and inert to hide non-modal elements
Although we use the aria-modal attribute on the dialog content, this does not fully hide the
background content from all assistive technologies - notably speech recognition tools. This PR
introduces a new custom hook that adds the `aria-hidden="true"` and `inert` attributes to all
containers of non-modal content, to further mark them as non-active for assistive technologies. The
attributes are removed - or returned to any previous values they had - when the modal is closed. The
new hook also determines whether a modal is the top modal or not, which was previously done inside
each individual component, in order to centralise and simplify this logic.

fix #5920