fix(switch): ensure invariant is not thrown when loading and validation is set
the `loading` prop has been passed down from `Switch` to `FormField` to be used within an invariant
which throws a violation when any validation is set and `disabled` is true. This was thrown
previously, as `disabled` is set to true when `loading` is also true. By now passing `loading`
down, a clear distinction is created between the two props, ensuring the invariant cannot
be thrown when `loading` is true