fix: set rootDir to src in tsconfig
The `build:types` script ran as part of the release process is broken as it outputs the `.d.ts`
files inside `src` directories in the output folders. On investigation, this has happened because of
the addition of `.ts` files inside the new `playwright` folder which are not excluded from the ts
build and which cause tsc to default the `rootDir` compiler option to the root directory, rather
than the `src` folder as it needs to be. To fix this, `rootDir` has been explicitly set to `./src/`
in the build config, as well as the playwright ts files being excluded (without that tsc throws
errors about trying to build files outside of `rootDir`). To avoid breaking the local `type-check`
script which still needs to typecheck the playwright ts files, the new config has only been added
inside the `tsconfig-build.json`.