feat(definition-list): allow children to be wrapped
Removes children iteration which was constraining the ability to wrap children in
custom components. This update may result in minor layout changes due to how spacing
defaults are applied, it is recommended to use the existing spacing interfaces
surfaced on the `Dt` and `Dd` components to correct them.

fix #6216, fix #6243