feat(tile): refactor component to support more flexible composition of children
Surfaces a new `TileContent` component to support the same layout as previous `Tile` component.
Removes the iteration over the `children` to better support more flexibility in consumers’
implementations and fix issue where returning `null` in a child causes additional dividers to be
rendered.

BREAKING CHANGE: You will need to wrap children in `TileContent` to achieve the
layouts of the
previous component where they receive padding relative to that of the `Tile`
parent's.

fix #5760