feat(portrait, profile): align with design system and reduce code complexity
aligns components with the sage design system, to ensure consumers can use the component with
the expected styling and functionality. This includes the addition of a new dark background
variant on `Profile`, as well as ensuring the same variant is consistent on the sub-component
`Portrait`. This aim for consistency has also been applied to all use of colour and background
colour tokens, ensuring they are aligned with the design system, which also removes some colour
contrast issues which were present due to the use of previous colours which are no longer
recommended by the design system.

As well as styling changes, the `text` prop has been added, which gives consumers the
ability to pass some read-only text underneath where the user's email is typically rendered.
This is commonly used for phone numbers etc. but can be used for any other added information
or context needed for a user's profile. Also, in regard to the rendered email, this is now
rendered as the `Link` carbon component, with some additional colour changes to ensure proper
colour contrast is achieved when used in the aforementioned dark background variant and also
now has a mailto href which is standard for emails in user profiles.

Code complexity has also been greatly reduced throughout both components, and their sub-
components. This has greatly improved code simplicity and consistency, particularly when
initials are rendered in `Portrait`, a new approach has been used which renders a div
instead of an image, eradicating some pixel abnormalities when the prop was used.

fix #6191