fix(pod): make edit button an HTML button rather than an anchor tag
Fix accessibility by ensuring the edit button is an actual button, rather than an anchor tag with no
href. This makes it possible to interact with the button with the keyboard as well as with a
screenreader.

BREAKING CHANGE: The onEdit prop is now restricted to being an event handler function - the options
for it to be a string or an object have been removed.