fix(navigation-bar): stop ResizeObserver errors being thrown on render
Whenever a NavigationBar component (or GlobalHeader, which renders NavigationBar) is rendered, an
error is raised "ResizeObserver loop completed with undelivered notifications". While not actually a
problem, and not even being reported to the browser console except on Safari, it triggers window
error handlers, which leads to problems in some environments that consume Carbon. The solution is to
take more care of the object that is passed to useResizeObserver - by using a React ref rather than
a custom object that is a new reference on each render, ResizeObservers are no longer destroyed and
recreated on each render.

fix #6259