feat(select): add selectionConfirmed property to event emitted on change
Adds a `CustomSelectChangeEvent` type which defines the shape of the new change event emitted when
the input value changes, backwards compatibility is maintained so it is still possible to continue
to use React's ChangeEvent. Within the new custom event there is a `selectionConfirmed` property
that indicates when a user has pressed `Enter` or clicked on a given `Option`.

fix #6330