fix(menu-full-screen): fix keys in rendering of children
Currently the component gives a warning in the console for the list of children not having keys. And
due to the lack of keys, complex situations involving a child being added or removed while another
changes state are not handled correctly. This has been fixed by using React's build in toArray
utility which automatically handles the keys.