fix: ensure any auto-focused element is focused first when a child of a focus trapped component
Ensures that when a Carbon input has the `autoFocus` prop set on it and rendered as a child of a
focus trapped parent (`Dialog`, `DialogFullScreen`, `Sidebar`, `Confirm` and so on) is focused first
when rendered. Also ensures that any elements with tabIndex of -1 are ignored by the tab guards,
rendered as part of the FocusTrap, when selecting which element to focus next.

fix #6384