fix(menu-segment-title): ensure segment titles are rendered as part of sublist with associated items
Refactors `MenuSegmentTitle` component to recieve associated list items as children. This ensures
that the heading and list items are held within a sublist so that the heading is treated correctly
by voiceover software. Updates the title element to support the `as` prop to allow it to be set to
any valid heading element but defaults to `h2` if no value is passed. Removes `display:
inline-block` from list items and makes main menu list a flexed layout.

BREAKING CHANGE: You will need to pass any associated list items as children of the
`MenuSegmentTitle` so they are rendered in a sublist with the segment heading

fix #6010