fix(select): ensure re-opening the select resets scroll position to the top
Updates the react-virtual dependency, which was directly responsible for the bug. The new version
makes the virtualiser return 0 items whenever the select is closed - this change in behaviour
resulted in some problems after the update, particularly with tests, which have been addressed. An
accessibility error with long virtualised lists, where selecting one option and scrolling away from
it so that the selected option is no longer in the DOM, resulting in an invalid
aria-activedescendant attribute, has also been addressed.

fix #6399