fix(input): ensure focus styles removed when becoming inert
Our input components manage focus styles using react state rather than CSS, due to limitations in
their implementation. This state is changed in the input's focus and blur events. However when
opening a modal, we add the `inert` attribute to all parts of the DOM tree outside the modal - this
causes the input to lose focus, as inert elements are not supposed to receive focus in any
descendants. In this situation, not all browsers fire the `blur` event on the input that's losing
focus - in particular, Firefox does not. This results in the focus style remaining on the
now-unfocused input. To fix this we need to manually call the `blur` method on any currently-focused
element before adding the `inert` attribute.

fix #6415