feat(card): ensure card action can be triggered by correct key presses when it is a button or link
Adds `onClick` and `href` props to component. When an `href` prop is passed an `a`nchor element is
rendered and the Card behaves like a link clicking or pressing the enter key when focused will
trigger the action. When an `onClick` prop is passed with no `href` then a `button` element is
rendered, clicking or hitting enter/space key will trigger the action. Passing the `draggable` prop
will negate both the `href` and `onClick` props. Surfaces a `footer` props for passing `CardFooter`
to avoid potential accessibility issues.

BREAKING CHANGE: Removes `action` and `interactive` props. Renames `cardWidth` prop to `width`.

fix #4800