fix(select): stop list scrolling to selected option unless it has changed
SelectList has a useEffect that sets the option to highlight and automatically scrolls to it. Due to
unavoidable depency chains, this effect runs whenever the option children are updated - including
when new children are added via lazy-loading, causing the scroll position to jump back to the
selected option, which is unwanted. The fix pulls the scroll functionality out into a separate
effect that only runs when the index of the selected option has actually changed.

fix #6461