feat: move Polish translation files into an internal directory
The Carbon team no longer has any native Polish speaking members so there is no guarantee of the
accuracy of the translations in this file. Therefore, it has been moved to an `internal` directory
as it should not be used outside of our storybook documentation.

BREAKING CHANGE: The `pl-pl` translation files have been moved to an internal directory