fix(select): maintain option ids according to key when children count changes
Currently Select assigns a random guid as ID to each of its option children if no ids are provided
explicitly, and regenerates this list of ids whenever the number of children changes. This results
in unexpected behaviour when a new option is dynamically added to the start of the list on
selection, as then the ids are regenerated and a critical axe error results from the
aria-activedescendant property no longer referencing a valid child ID. To fix this, guids are now
generated in Option itself (rather than SelectList as before) when the prop is not provided, and it
is up to consumer to set keys on the list of option children to ensure individual child components
maintain their identity correctly across renders even as children are added or removed.

fix #6378