fix(pages): maintain height of content when page slides in
Currently, when moving from a page with a small amount of content to one with more content, the
height of the new page content is truncated (with a scrollbar) and suddenly "jumps" to the new
height when the transition is completed - a rather unexpected effect. This is due to the absolute
position on the page content as it slides in, which stops its height being taken into account when
calculating the parent's height. The fix is to remove this absolute positioning (moving it up to the
page wrapper so that the slide effect still works).

fix #6492