feat(portal): add inertOptOut prop to keep the content interactable when a modal opens
Recent versions of Carbon mark all outside elements with the `inert` and `aria-hidden` attributes
when a modal component is opened. This makes it so that no outside content can be clicked on, or
interacted with using assistive technologies such as screenreaders - and is quite deliberate because
a modal usually means that nothing else in the document should be functional. However in order to
cater for special cases where a small area of content is still intended to be interacted with, this
commit introduces a new prop `inertOptOut` on the `Portal` component, which opts out of this
behaviour. If you wrap the important content in a `Portal` with the `inertOptOut` prop, this will
keep it clickable (and appearing visually on top of the modal) even when a modal is open.