feat(sort): add sortIcon prop to ensure sorting icons are customizable
the `sortIcon` prop has been added to provide consumers with the ability to
bypass the set icons for sorting up and down. The prop will accept any valid node,
however consumers are encouraged to pass the Icon component. From there, consumers
have complete control of the type of icon which is displayed, as well as further
control around the icons appearance and presentation

fix #6491