refactor(filterable-select): make story controlled and improve type-safety
The selectionConfirmed story uses an uncontrolled select, which is deprecated. When changed to be
controlled, it crashes with an invariant error when filtering with no match, due to types being
inferred as `any` - this has been fixed with more type annotations and a default value of an empty
string when there is no match.