feat(loader-spinner): add new component
Creates a new carbon component to aid consumers in providing end users with
transparency around wait times. Consumers can use either use a tracked or
un-tracked animation to display either predictable or unpredictable wait times,
change the components appearance with multiple variants, extend animation time
and conditionally render the components loading label.

Due to the nature of the component, measures have been taken to ensure accessibility;
the component itself serves as a polite live region to ensure that the component is
announced to assistive technologies upon rendering. If consumers choose not to
render a visual label to adhere to design specifications, a visually hidden label
is rendered instead to ensure that the presence of the component upon rendering is
still announced. Finally, if at any time the end user prefers reduced motion,
the spinner is not rendered, and only a loading label is displayed in its place