fix(select): objects as value fix
When values for options in the `Select` component are set to objects, select-list is not able to
compare object values and does not highlight the selected option as it does not find it in the
children values. Added a shallow equality check to be able to compare property values instead of
object references.

fix #6600