build: remove proptypes from production bundles
Carbon components are released with `propTypes` definitions, automatically generated from the
Typescript types, which are in some cases very large (eg Box has over 3500 lines of proptypes in the
npm package). These can be useful in development for teams not using typescript, but are not needed
in production (and proptypes warnings are turned off by React in production builds anyway). To
reduce the size of consumers' builds, this change wraps the generated proptype definitions in a
conditional check on `process.env.NODE_ENV`, so that they will still be there in development but
will be stripped out of production builds by the bundler.

BREAKING CHANGE: Carbon components will no longer expose `propTypes` in production builds - any
functionality that relies on reading the `propTypes` of a component will break