fix(loader-spinner): allow components label to extend longer than the svg wrapper
removes the fixed width on the components svg wrapper and label to ensure the component
centers when rendered. Due to these changes the components label can now extend longer
than the svg wrappers width. A varied margin-top is also now applied to the label on
all sizes above extra-small, to ensure there is a correct amount of space between the
label and svg wrapper. Finally, the labels text has also now been aligned to the center,
so if any text does wrap due to container constraints, it will center correctly

fix #6702