feat(step-flow): allow component title to be compositional
the sub component `StepFlowTitle` has been created to enable consumers to use a compositional
approach, and render custom title content. Consumers can now pass a node to the `title` prop
and pass any relevant elements to achieve designs and meet requirements - as well as the
`StepFlowTitle` within that node, to ensure the component still functions correctly.

Within the `StepFlowTitle` sub component, consumers can pass the `titleString` prop to override
the text children, `titleVariant` to change the heading level between h1 and h2 and the
`screenReaderOnlyTitle` to provide an alternative screen reader only title. This is useful
if added information or context needs to be added to the title, especially if other elements
have now been rendered in the node passed to the `title` prop.

fix #6732