feat(preview): add `shape` and `disableAnimation` props
Adds `shape` prop which allows consumers to set the shape of a Preview to "text", "rectangle",
"rectangle-round" and "circle". These shape variations have default dimensions however the existing
`width` and `height` props can be used to change these. The "circle" shape will use the set `height`
as its diameter, ignoring the `width`. The `lines` prop can still be used to render the specified
amount of Previews. Also adds the `disableAnimation` prop to disable the loading animation, this can
be set manually or will automatically be true when prefer reduce-motion is enabled.