fix(toast): ensure closed toasts do not trigger axe warnings
When a Toast's content is only defined on open, its accessible name cannot be computed and is
"empty" while closed. This causes an axe issue when more than one Toast is rendered on a page with
undefined content as their accessible name is not "unique". This fix ensures that the region is
aria-hidden while closed and aria-labelledby is only set when the Toast is opened.