fix(navigation-bar): move context provider to not wrap parent
The FixedNavigationBarContextProvider, which uses a ResizeObserver to respond to size changes on the
navbar element, currently wraps that same element. This can result in unexpected ResizeObserver
errors on initial render, when a media query is being used to determine layout. The context provider
only needs to wrap the navbar's children, and by moving it down the React tree to only wrap the
children and not the parent, the error goes away.