fix(action-popover): ensure that opening using the up arrow focuses last element in the menu
Currently we have a bug whereby when a user presses the up arrow key, the first focusable item in
the menu is focused. This appears to be a bug as we have code that specifically handles this
behaviour but is overridden elsewhere in the code. This fix ensures that the last focusable item is
focused.

fixes #6826