fix(switch): fix screen reader announcements with new validation
This fix amends the screen reader announcements with voiceOver in Safari, preventing "and x more
items" to be announced. Sets the input's accessible name as the label and accessible description to
the hint text and validation message if provided. This fix also addresses issue where required star
or (optional) did not appear on the label when set with the neww validation.

fix #6901