feat: introduce support for react v18
BREAKING CHANGE: Temporary incompatibility with React Strict Mode. In React 18, Strict Mode
now renders components twice in development, which causes unexpected behaviour in several Carbon
components.

We are actively working on a fix, which will be included in an upcoming release. In the meantime,
if you encounter issues, you can either disable Strict Mode globally, or apply it selectively to
specific parts of your application.