test(filterable-select): remove tests asserting wrapper component behaviour
Removed tests focused on the lazy loading behaviour managed by the test wrapper. This behaviour
is not a built-in feature of `FilterableSelect` so the tests did not reflect its out-of-box
functionality. The test suite now focuses on verifying `FilterableSelect`'s native behaviour
based on its props and interactions with its rendered UI.