feat(text-editor): replaces draftjs with lexical
The draftjs package is no longer maintained; a decision has been reached to rewrite the TextEditor
component using Lexical

BREAKING CHANGE: Whilst the intention is to implement a straight-up replacement and maintain as much
functionality as possible, there may be differences that customers need to be made aware of. These
will be communicated separately as necessary.