chore: check for window existence
I should probably write a formal ticket for this,
but here's the commit for now anyways.

This PR supports applications to use SSR with
React 18 by checking for window
object availability in places
other than useEffects.

I needed these changes to make our admin portal
+ react-router@7's new routing API work.

There should probably be more robust typing or
linting available to check for window
if this PR is generally accepted
since it's easy for a dev to break the
SSR build by accident.