test(playwright): remove lint rule encouraging flaky behaviour in helper function
The ESLint rule `no-await-in-loop` encourages running all asynchronous tasks in a loop
concurrently and waiting for them to all complete. However, this assumes each iteration
is independent, which is not the case for Playwright actions, such as key presses and clicks,
which may depend on previous actions completing first. Following this rule could lead to
test flakiness.