fix(menu, multi-action-button, split-button): remove focus first list item on submenu open
Focus is now maintained on the button that toggles the menu on open, instead of moving to the first
item within the list. Enter/Space should toggle the menu closed, MultiActionButton and
SplitButton menus can also be closed on click. Support to open the submenu with ArrowUp has been
removed.

fix #7264