feat(storybook): introduce Play functions
Introduce automated interactions into Storybook. By default, these interactions auto-play which can
be problematic for users who prefer reduced motion, so a control switch has been added to toggle
this functionality. It also uses media queries to check if "prefers-reduced-motion" has been set in
the OS, so that it can be off by default without the need for user input.