feat(form): add stickyFooterVariant prop
adds the stickyFooterVariant prop which gives consumer control over the background variant of the
sticky footer. The default is "light", which is the default light colour which was already available
with the component. However, a new "grey" variant has been added which applies a light grey
background colour. If the "grey" variant is added, an additional border-top is applied to ensure
there is enough contrast if other background variants are used elsewhere