feat(dialog): merge dialog-fullscreen and dialog into single component
There is a lot of duplicated code between Dialog and DialogFullscreen, yet the only real difference
is whether or not the dialog takes up the full screen or not. Both components are therefore viable
candidates to be merged together into a single Dialog component, with optional fullscreen flag

BREAKING CHANGE: DialogFullscreen has been deprecated and will soon be removed. Use Dialog with the
fullscreen prop instead.