feat(popover-container): add variant prop
adds the `variant` prop, when set to "quick-action" the popover container
will apply a green background colour and change the close icon to white.
This variant is to signify that the popover container is now used and will
contain elements which achieve a quick action.

Border radius will be removed from the top of the popover, and all offset is
removed, this makes the popover container appear more connected to the
trigger element above.

When inside a `MenuItem and ``renderOpenComponent` is used with a `Button`
this green background will also be applied. This shows the button and the
popover container are linked, and both share the same purpose of triggering
quick actions.