fix(responsive-vertical-menu): ensure secondary menu is a child of primary
Moves secondary menu element to be a child of the primary menu instead of a sibling to ensure
correct screen reader behaviour. Also fixes issue where if a tertiary menu item with children was
the last item of the secondary menu, its children are not accessible via keyboard navigation.

`ResponsiveVerticalMenuProvider` had been deprecated, consumers can now use
`ResponsiveVerticalMenu` directly without the need to wrap it in a provider.