feat: ensure popover based components cannot remain open when an adaptive sidebar renders as a modal
fixes an issue where the adaptive sidebar can be conditionally rendered as a modal and popover based
components remain open. This issue only exists due to the adaptive sidebar potentially being a div
or a modal conditionally. In other patterns an element will typically be pressed or clicked to open
a modal which results in all open popovers being closed