fix(reponsive-vertical-menu-item): item rendering and onClick behaviour consistency
when the responsive vertical menu item is rendered as an anchor element the onClick
behaviour is now consistently aligned with native HTML behaviour. The responsive
vertical menu item now also supports being rendered as a button if no href is passed
and an onClick is passed. This means the onClick can be triggered on space & enter
key press and on a click event for specific use cases

fix #7575

BREAKING CHANGE: the onChange callback on responsive vertical menu item
now handles for click/keyboard events on both anchor and button elements. If you
have created a custom click handler with a typed event this will now cause a type error.
This is due to a contravariance limitation, the type will not be able to accept a broader
event union. We recommend importing the new ResponsiveVerticalMenuItemClickEvent type
and using this in your custom click handler