fix(file-input): add download attribute to completed state link
when the upload status of the file input is completed, a link is rendered. This now supports the
native download attribute, if the link is suitable for download it will trigger a download event in
the browser when pressed. This ensures behaviour is more consistent between the major browsers

fix #7633